feat: event for batch operation
