+++ zrcs/zrc-6.md
@@ -341,6 +341,12 @@ Mints tokens and transfers them to `to_list`.
 | ------ | ------------------------ | ----------------------------------- | ------------------- |
 | `_tag` | `ZRC6_BatchMintCallback` | Provide the sender with the result. |                     |
 
+**Events:**
+
+|              | Name        | Description                       | Event Parameters                                                |
+| ------------ | ----------- | --------------------------------- | --------------------------------------------------------------- |
+| `_eventname` | `BatchMint` | Multiple tokens have been minted. | `to_list` : `List ByStr20`<br/>Addresses of the token recipient |
+
 #### 8. `Burn` (Optional)
 
 Destroys `token_id`.
@@ -564,6 +570,12 @@ Transfers multiple `token_id` to multiple `to`.
 |        | Name                             | Description                         | Callback Parameters |
 | ------ | -------------------------------- | ----------------------------------- | ------------------- |
 | `_tag` | `ZRC6_BatchTransferFromCallback` | Provide the sender with the result. |                     |
+
+**Events:**
+
+|              | Name                | Description                            | Event Parameters                                                                                                                                                                                                                 |
+| ------------ | ------------------- | -------------------------------------- | -------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `_eventname` | `BatchTransferFrom` | Multiple tokens have been transferred. | `to_token_id_pair_list`: `List (Pair ByStr20 Uint256)` <br/><br/> List of Pair (`to`, `token_id`)<br/><br/><ul><li>`to` : `ByStr20`<br/>Address of a recipient</li><li>`token_id` : `Uint256`<br/>Unique ID of a token</li></ul> |
 
 #### 16. `SetContractOwnershipRecipient` (Optional)
